feat: Add Darkness spell to Drow race
This commit adds the Darkness spell to the Drow race in the `races.py` file. The `Darkness` spell allows Drow characters to evoke a magical darkness that darkens an area. This spell is added to the `initialize_race_features()` method of the Drow race.

This change is necessary to enhance the gameplay experience for Drow characters by giving them access to the Darkness spell.