feat: Add ElfFire spell to Elf race
This commit adds the ElfFire spell to the Elf race in the `races.py` file. The `ElfFire` spell allows Elf characters to hurl a mote of fire. This spell is added to the `initialize_race_features()` method of the Elf race.

This change is necessary to enhance the gameplay experience for Elf characters by giving them access to the ElfFire spell.